requirements.txt
django
django-parler
-e git+https://github.com/tback/django-hvad.git@dev/hvad2#egg=django-hvad
nece
django-modeltranslation==0.13-beta1
django-i18nfield
psycopg2
django-extensions
django-modeltrans
django-translated-fields
faker

